Remove client from all except handlers.js
